requirements.txt
Django==2.0.3
numpy==1.14.1
Pillow==5.0.0
pytz==2018.3
scikit-learn==0.19.1
scipy==1.0.0
keras
tensorflow==1.8.0
